autotool improvements
