+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/world/ModuleAutoTool.kt
@@ -19,12 +19,14 @@
 package net.ccbluex.liquidbounce.features.module.modules.world
 
 import it.unimi.dsi.fastutil.ints.IntObjectImmutablePair
+import net.ccbluex.liquidbounce.config.types.Choice
+import net.ccbluex.liquidbounce.config.types.ChoiceConfigurable
 import net.ccbluex.liquidbounce.event.events.BlockBreakingProgressEvent
 import net.ccbluex.liquidbounce.event.handler
 import net.ccbluex.liquidbounce.features.module.Category
 import net.ccbluex.liquidbounce.features.module.ClientModule
 import net.ccbluex.liquidbounce.utils.client.SilentHotbar
-import net.ccbluex.liquidbounce.utils.item.isNothing
+import net.ccbluex.liquidbounce.utils.inventory.findBestToolToMineBlock
 import net.minecraft.block.BlockState
 import net.minecraft.entity.player.PlayerInventory
 import net.minecraft.item.ItemStack
@@ -36,20 +38,37 @@ import net.minecraft.util.math.BlockPos
  * Automatically chooses the best tool in your inventory to mine a block.
  */
 object ModuleAutoTool : ClientModule("AutoTool", Category.WORLD) {
+    val toolSelector =
+        choices(
+            "ToolSelector",
+            DynamicSelectMode,
+            arrayOf(DynamicSelectMode, StaticSelectMode)
+        )
 
-    // Ignore items with low durability
-    private val ignoreDurability by boolean("IgnoreDurability", false)
+    sealed class ToolSelectorMode(name: String) : Choice(name) {
+        abstract fun getTool(inventory: PlayerInventory, blockState: BlockState?): IntObjectImmutablePair<ItemStack>?
+    }
 
-    // Automatic search for the best weapon
-    private val search by boolean("Search", true)
+    private object DynamicSelectMode : ToolSelectorMode("Dynamic") {
+        override val parent: ChoiceConfigurable<*>
+            get() = toolSelector
 
-    /* Slot with the best tool
-     * Useful if the tool has special effects
-     * cannot be determined
-     *
-     * NOTE: option [search] must be disabled
-     */
-    private val slot by int("Slot", 0, 0..8)
+        private val ignoreDurability by boolean("IgnoreDurability", false)
+
+        override fun getTool(inventory: PlayerInventory, blockState: BlockState?) =
+            inventory.findBestToolToMineBlock(blockState, ignoreDurability)
+    }
+
+    private object StaticSelectMode : ToolSelectorMode("Static") {
+        override val parent: ChoiceConfigurable<*>
+            get() = toolSelector
+
+        private val slot by int("Slot", 0, 0..8)
+
+        override fun getTool(inventory: PlayerInventory, blockState: BlockState?): IntObjectImmutablePair<ItemStack> {
+            return IntObjectImmutablePair(slot, inventory.getStack(slot))
+        }
+    }
 
     private val swapPreviousDelay by int("SwapPreviousDelay", 20, 1..100, "ticks")
 
@@ -67,32 +86,8 @@ object ModuleAutoTool : ClientModule("AutoTool", Category.WORLD) {
 
         val blockState = world.getBlockState(pos)
         val inventory = player.inventory
-        val index = getTool(inventory, blockState)?.firstInt() ?: return
+        val index = toolSelector.activeChoice.getTool(inventory, blockState)?.firstInt() ?: return
         SilentHotbar.selectSlotSilently(this, index, swapPreviousDelay)
-    }
-
-    fun getTool(inventory: PlayerInventory, blockState: BlockState?): IntObjectImmutablePair<ItemStack>? {
-        if (search || !running) {
-            val (hotbarSlot, stack) =
-                (0..8).map {
-                    it to inventory.getStack(it)
-                }.filter { (_, stack) ->
-                    val durabilityCheck = (stack.damage < (stack.maxDamage - 2) || ignoreDurability)
-                    (stack.isNothing() || (!player.isCreative && durabilityCheck))
-                }.maxByOrNull { (_, stack) ->
-                    stack.getMiningSpeedMultiplier(blockState)
-                } ?: return null
-
-            val miningSpeedMultiplier = stack.getMiningSpeedMultiplier(blockState)
-
-            // The current slot already matches the best
-            if (miningSpeedMultiplier == player.inventory.mainHandStack.getMiningSpeedMultiplier(blockState)) {
-                return null
-            }
-            return IntObjectImmutablePair(hotbarSlot, stack)
-        } else {
-            return IntObjectImmutablePair(slot, inventory.getStack(slot))
-        }
     }
 
 }
